move notification schedule out of keystore
